Relates to Bug [ 973294 ] NoSuchElementException in URI queues halts crawling
Still not a fix for the underlying problem. Simply a fix for the workaround that previous commit was supposed to implement. That is
now when the bug is encountered the relevant queue is FROZEN, removed from the ready queues list and the thread can retry 
getting a CURI from another queue (assuming one is available).