Fix for [ 989816 ] Specification of default CharSequence charset Fix submitted by Tom Emerson, tree at lists dot sourceforge dot net Here is his comment on the patch:
'...adds support for specifying the "default" charset used when crawling
pages that do not specify on in the HTTP Content-Type response header.

...

NOTE: the default value for the setting is the default charset used by
the Commons HTTP client, which is ISO-8859-1. Previous discussion
centered on using the file.encoding system property, but I don't think
this is a good idea. The code is structured such that changing this
discussion will be easily done.

Also, I noticed that option settings for the proxy hosts were slightly
off: the code looked like they were supposed to be marked as expert
settings, but the calls to addElementToDefinitions() did not store the
returned element. I fixed this.'

* src/java/org/archive/crawler/fetcher/FetchHTTP.java
    Fix to make proxy setting expert setting (Was setting 'expert' on
    the option setup just previous because wasn't catching returned
    option pointer).
    (ATTR_DEFAULT_ENCODING, setCharacterEncoding): Added.