More from Oskar on '[ 1232339 ] [DomainSensitiveFrontier] Toggle count overall/per host':
"....I have (AGAIN) made some changes to the DSF.  It turned out that the
Danes had about 10 000 domains that they wanted to set max-docs for and
didn't find it too fun to have to do that many overrrides.
So I added a new mode that acts like count-per-host, but on the domain-level,
i.e., you set max-docs once and it downloads that many docs from each domain
in total. So now you choose between these three modes how you want it to act.
As always see if it seems to do any good and feel free to change whatever you
feel like."

I did not test the new feature other than make sure it makes
sense in the UI

* src/java/org/archive/crawler/frontier/DomainSensitiveFrontier.java
    Added new counter-mode option: count-per-domain.