HER-1136 fix.
* heritrix/src/main/java/org/archive/crawler/framework/CrawlJobManagerConfig.java
New class that encapsulates configuration information for
CrawlJobManager, as it now takes five parameters.  Noteworthy are the
mutators setJobsDirectory and setProfilesDirectory.

* heritrix/src/main/java/org/archive/crawler/framework/CrawlJobManagerImpl.java
(rootDir): Deleted; was used to provide root of both profile and jobs dirs.
(jobsDir): New variable for keeping track of the jobs directory.
(profileDir): New variable for keeping track of the profiles directory.
(<init>): Now takes a CrawlJobManagerConfig and uses that to set instance
variables.

* heritrix/src/main/java/org/archive/crawler/Heritrix.java
(main): Overhauled to use Apache Commons CLI again.  Now takes 3 optional 
command line arguments: -j for setting the jobsDir, -p for the profileDir,
and -h for the heritrix.properties file.