Made logs available via new UI.
* heritrix/src/main/java/org/archive/crawler/CrawlJobManager.java
* heritrix/src/main/java/org/archive/crawler/CrawlJobManagerImpl.java
(openLogs) Added a method to create a LogRemoteAccess for the current job.

* heritrix/src/main/java/org/archive/crawler/util/LogRemoteAccess.java
* heritrix/src/main/java/org/archive/crawler/util/LogRemoteAccessImpl.java
Added a JMX bean interface to provide remote access to logs and the relevant implementation thereof.
Enables the 4 previous access methods (tail, by regexpr, by timestamp and by linenumber).

* heritrix/src/main/java/org/archive/crawler/util/Logs.java
An enum for existing logs. 
Attempts to access logs not covered by this enum will result in an error.

* webui/src/main/java/org/archive/crawler/webui/Log.java
* webui/src/main/webapps/logs/do_show_log.jsp
* webui/src/main/webapps/logs/page_show_log.jsp
Added new page to UI to display logs.

* webui/src/main/webapps/crawler_area/page_crawler.jsp
Added links to do_show_log.jsp for ongoing and completed jobs.