New webui pages for associating SURT prefixes with sheets, and for previewing the configuration that will be applied to a URL.
* webui/src/main/java/org/archive/crawler/webui/Sheets.java
(showAssociate)
(associate)
(showConfig): New action methods for making/checking associations.

* webui/src/test/java/org/archive/crawler/webui/WebUIJUnit.java
(testWebui): Extended the test so that it creates a new single sheet
'foo', overrides a setting in 'foo', associates 'foo' with a SURT
prefix, and previews the configuration for a URL with that prefix.

* commons/src/main/java/org/archive/settings/SheetManager.java
(findConfig): Changed return type to Sheet so it can be passed to PathLister.
(findConfigNames): New method that just returns the names of the sheets
that would be used to mine settings for a particular URL.

* commons/src/main/java/org/archive/settings/SheetBundle.java
(resolveNormal): Fixed bug that prevented bundles from working properly
in an offline sheet manager.

* commons/src/main/java/org/archive/settings/jmx/JMXSheetManager.java
* commons/src/main/java/org/archive/settings/jmx/JMXSheetManagerImpl.java
(findConfig)
(findConfigNames): Exposing additional methods of SheetManager to JMX.

* webui/src/main/webapp/sheets/do_associate.jsp
* webui/src/main/webapp/sheets/do_show_associate.jsp
* webui/src/main/webapp/sheets/do_show_config.jsp
New action pages.

* webui/src/main/webapp/sheets/page_associations.jsp
New HTML form for entering URLs to associate with a sheet.

* webui/src/main/webapp/sheets/page_config_names.jsp
New HTML page for displaying the sheets that would be applied against a URL.

* webui/src/main/webapp/sheets/page_sheet_detail.jsp
Modified this page to more closely resemble page_sheet_editor.

* webui/src/main/webapp/sheets/page_sheets.jsp
Added associate/disassociate links, and added new form for viewing
settings for a particular URL.