Fixing the build.  Moving the tarball creation up to parent project worked during install phase, but broke during deploy phase due to this bug:
http://www.mail-archive.com/users@maven.apache.org/msg55883.html

Fixed by using a moduleSet in the assembly descriptor instead of relying
on fileSets to provide the /lib contents.  That eliminated the need for the
heritrix subproject to have its own assembly descriptor.

* src/main/assembly/dist.xml
Use moduleSet instead of fileSet for populating /lib in tarball.

* heritrix/pom.xml
Remove maven-assembly-plugin, as it is no longer used/desirable.