Work for HER-1364: Referential woe
Problem was actually broader than originally reported.  If a module were
referenced by other modules, and the referant module changed, then the
references would still point to the old module. 

My approach was to modify SingleSheet, SettingsMap and SettingList so
that they remember where a module was first defined.  If the module is
swapped out at that site, then all references to that module are
instantly updated with the new module.

This process revealed many other little bugs that got fixed along way. 
However, see HER-1364 for the five minor things still outstanding here,
and see HER-1312 for thoughts on making this gnarly, gnarly code
simpler.

* commons/src/main/java/org/archive/settings/ModuleInfo.java
New structure for keeping track of whether or not this is the first site
(path) where a module was defined.
(Holder): Introduces a level of indirection between a setting value and
the actual module.

* commons/src/main/java/org/archive/settings/SingleSheet.java
(settings): Now a WeakHashMap.  If a key maps to a module value, then
the actual value is now a ModuleInfo, not the module itself.
(holders): New structure for mapping a module to the holder for that module.
(setModuleValue): New method for figuring out whether or not to change
just one reference to a module, or all of them.
(set2): Delegate to setModuleValue if the key is a module type.
(duplicate): Deep copy the holders map as well.

* commons/src/main/java/org/archive/settings/SettingsMap.java
(delegate): Lost its type safety.  If the element type of the map is a
module type, then the delegate actually maps to ModuleInfo and not the
module type itself.
(toActualObject): Converts ModuleInfo to module instance if necessary.
(put): Use SingleSheet.setModuleValue to see if we need to change one
reference or all of them.

* commons/src/main/java/org/archive/settings/SettingsList.java
(delegate): Lost its type safety.  If the element type of the list is a
module type, then the delegate actually holds ModuleInfo and not the
module type itself.
(toActualObject): Converts ModuleInfo to module instance if necessary.
(put): Use SingleSheet.setModuleValue to see if we need to change one
reference or all of them.

* commons/src/main/java/org/archive/settings/UnmodifiableTypedList.java
* commons/src/main/java/org/archive/settings/UnmodifiableTypedMap.java
Can now work on generic Map and Lists, not just TypedMap and TypedList.

* commons/src/main/java/org/archive/settings/UnspecifiedSheet.java
Now returns UnmodifiableTypedList and UnmodifiableTypedMap instead of
SettingsMap and SettingsList.  Necessary because SettingsMap and
SettingsList are now tightly tied to SingleSheet.

* commons/src/main/java/org/archive/settings/Duplicator.java
Now duplicates ModuleInfo instances correctly.

* commons/src/test/java/org/archive/settings/path/PathChangerTest.java
Added more unit tests to ensure this all works as advertised.

* dist/src/main/conf/jobs/*/sheets/global.sheet
Moved uri-canonicalization-rules list to the Frontier key that actually
needs them, instead of these living at root.  See note at HER-1364.

* commons/src/main/java/org/archive/settings/SheetMap.java
Deleted unused class.

* commons/src/test/java/org/archive/settings/SheetMapTest.java
And deleted its really long unit test.